refactor: 순수함수가 아닌 reducer 수정
